feat(back): implemented rating imc CRUD

+++ backend/src/modules/user/useCases/deleteUser/DeleteUserUseCase.ts
@@ -1,23 +1,25 @@
 import { inject, injectable } from 'tsyringe'
 
 import { IUserRepository } from '@modules/user/repositories/IUserRepository'
+import { IRatingImcRepository } from '@modules/rating/repositories/IRatingImcRepository'
 import { AppError } from '@shared/errors/AppError'
 
 @injectable()
 export class DeleteUserUseCase {
   constructor(
     @inject('UserRepository')
-    private userRepository: IUserRepository
+    private userRepository: IUserRepository,
+    @inject('RatingImcRepository')
+    private ratingImcRepository: IRatingImcRepository,
   ) { }
 
   async execute(id: string): Promise<void> {
     const user = await this.userRepository.findById(id)
     if (!user) throw new AppError('Usuário não existe na base de dados', 404)
 
-    /**
-     * TODO: Buscar se o usuário tem avaliação de imc ou foi o responsável por incluir a informação de uma avaliação,
-     * se sim não permite a exclusão
-     */
+    const hasStudentHistoric = await this.ratingImcRepository.find({ user_student_id: user.id, skip: 0, limit: 1 })
+    const hasRatingHistoric = await this.ratingImcRepository.find({ user_rating_id: user.id, skip: 0, limit: 1 })
+    if (hasStudentHistoric || hasRatingHistoric) throw new AppError('Não é possível excluir o usuário, ele possui histórico de avaliações')
 
     await this.userRepository.delete(id)
   }
